Agent Server Interface for v2 (#37)
* Reenable openapi,json endpoint

In order to test feature parity with V1 for Workroom client, I want to create a new agent client for Workroom and see what I break.
For this reason, I need to generate OpenAPI spec.

I don't know why the openapi endpoint was disabled in first place, but I would expose it because it is useful in several scenarios
1. generate client code;
2. allow Ace to know the API version used by existing agent servers;

* new agent server client to test what we break in workroom

* restore

* generate schema and spec patching path with server url

* include only lib in package

---------

Co-authored-by: Marco Stenico <marco@sema4.ai>