GPT-796: Studio/Workroom Integration Round 3 (#63)
Getting this in since it has tools fixes and package endpoint. Future work to think harder on how agent-cli/spec live in v2 and how we centralize/de-dupe some of that code. (Server or Core or new folder? etc.)

After this merge, we are in a good place for an alpha pre-release I think (mostly, probably a few more minor things we'll find).