[ENG-14] Support for Dynamic OAuth2 for MCP (login/logout/refresh/status APIs) (#1837)
Provides OAuth2 support to the agent-server backend (no UI work done so far).

In practice, given that an mcp server at an address such as:
http://localhost:8080/mcp

One can log-into it by using:
http://localhost:58885/api/v2/oauth2/login?mcp_server_url=http%3A%2F%2Flocalhost%3A8080%2Fmcp

And then get the status using:
http://localhost:58885/api/v2/oauth2/status?mcp_server_url=http%3A%2F%2Flocalhost%3A8080%2Fmcp

Note that these calls will store the token/get its status based on the current user (so, it works for Studio directly as there's really no user management, but for the cloud we need to do the UI integration for the user to be the correct one).